Add endpoints to run Danesfield algorithm jobs
Add endpoints to run Danesfield algorithm jobs using Girder Worker.
Currently, two endpoints are exposed:

- /generate_point_cloud: Runs P3D on the specified images and places the
output point cloud files in an item in the "dataset" folder.

- /generate_dsm: Runs the Danesfield generate-dsm.py tool with the
specified point cloud file as input and writes an output file to the
"dataset" folder.

In case the job fails or is canceled, the placeholder output items that
the endpoints created are removed.

Metadata is recorded on the output items to track which algorithm
generated them. This mechanism is a work in progress that might be
useful for future work on a composite workflow.