Send notification on workflow error
Following a successful job, the workflow now advances to the next step
in an asynchronous handler. This avoids affecting the status of the
finished job.

If an error occurs while advancing the workflow, send a notification to
the client. The notification is of type 'danesfield_workflow_error' and
includes fields for 'step', 'previousStep', and 'message'.