Run workflow steps in parallel
Because there is a wall-clock time deadline within which to run the
complete processing pipeline, it's important to run as many steps in
parallel as possible, using multiple workers.

This commit changes how the workflow is configured to allow each step to
specify its dependencies. When advancing the workflow, the workflow
manager uses this information to determine which steps are able to run.
Multiple steps that have their dependencies satisifed may start running
at once.

Additionally, this commit reorganizes the code. The functions previously
in workflow_handlers.py are moved to individual steps; see the
workflow_steps folder. The algorithms are also refactored into
individual files.