Create distinct output folders for each run of the processing workflow
This commit makes the process endpoint create a new output folder. The
folder is named after the initial working set and a timestamp.

Previously, every run of the processing working used the same output
folder. This is confusing, because generated files may have name
collisions with existing files, and Girder adds numerical suffixes to
their names.