Improve event handling for composite steps
This commit changes how the job update event handler manages jobs in
composite steps.

Previously, there was a race condition where the celery
GroupResult would indicate that all the tasks had completed, but the
Girder job update event hadn't yet triggered for all the jobs. In this
case, the workflow manager stepSucceeded() method was called, but would
be called again when the next job update event arrived.

Now the workflow manager maintains and checks a count of the remaining
jobs. Both that count and the GroupResult status are considered when
determining whether a composite step is complete.