Added previous working set parameter to job initialization
If this parameter is provided, during job initialization the previous
working set's children working sets will be copied and link to the
current working set, and the steps used to generate these child
working sets will be flagged as complete.  This change should allow us
to resume processing of a failed run by providing the top level
working set id of the failed run.